Updated valid postion method to allow for arbitrary Y
Previously we required the player to spawn on the bottom row, but now we spawn elsewhere, which complicates finding a valid position. Additionally, we were not handling the case where the min radius might end up being invalid with the grid if the player is in the middle of the y axis.